+++ src/Form/ProjectType.php
@@ -4,6 +4,7 @@ namespace App\Form;
 
 use App\Entity\Project;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -12,16 +13,22 @@ class ProjectType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('name')
-            ->add('projectTrackerProjectUrl')
-            ->add('projectTrackerKey')
-            ->add('projectTrackerId')
-            ->add('include')
-            ->add('createdBy')
-            ->add('updatedBy')
-            ->add('createdAt')
-            ->add('updatedAt')
-            ->add('clients')
+            ->add('projectLeadName', null, [
+                'required' => true,
+                'attr' => ['class' => 'form-element'],
+                'label' => 'project.project_lead_name',
+                'label_attr' => ['class' => 'label'],
+                'row_attr' => ['class' => 'form-row'],
+                'help' => 'project.project_lead_name_helptext',
+            ])
+            ->add('projectLeadMail', EmailType::class, [
+                'required' => true,
+                'attr' => ['class' => 'form-element'],
+                'label' => 'project.project_lead_mail',
+                'label_attr' => ['class' => 'label'],
+                'row_attr' => ['class' => 'form-row'],
+                'help' => 'project.project_lead_mail_helptext',
+            ])
         ;
     }